Rename utilities directory.
