+++ lib/themes/dosomething/paraneue_dosomething/js/app.js
@@ -21,9 +21,9 @@ import HotModule from 'campaign/HotModule';
 import 'campaign/SchoolFinder';
 
 // Utilities
-import Share from 'utils/share';
-import ShowMore from 'utils/show-more';
-import 'utils/analytics';
+import Share from 'utilities/share';
+import ShowMore from 'utilities/show-more';
+import 'utilities/analytics';
 
 // Patterns
 import 'patterns/tiles';
