+++ lib/themes/dosomething/paraneue_dosomething/js/patterns/gallery.js
@@ -1,7 +1,7 @@
 import $ from 'jquery';
 import debounce from 'lodash/function/debounce';
-import Swapper from 'utils/Swapper';
-import Revealer from 'utils/Revealer';
+import Swapper from 'utilities/Swapper';
+import Revealer from 'utilities/Revealer';
 
 /**
  * Initialize galleries.
